backend/requirements.txt
fastapi[standard]
pydantic
pydantic-settings
sqlalchemy
asyncpg
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
openai
beautifulsoup4
requests
pypdf
httpx
pgvector
redis.asyncio
redis
